Return empty tuple for no selected files
This change makes the FileEdit in EXISTING_FILES mode return an empty tuple if no files are selected.

The ability to detect "no selected file" is important to prevent parsing files that don't exist. The FileEdit has three possibilities, either it always has a Path value, or it is nullable and the neutral value should be None, or it is in multiple file selection mode and according to the signature, the neutral element should be an empty tuple.

However, since `""`.split(some_delimiter)` always gives `[""]`, it currently returns a tuple with an empty path `Path(".")` which equals the home directory and is very impractical to distinguish from an intentionally selected path.